Research Conducted
I don't think it would be an effective use of my time to look at noise as their are already a lot of existing libraries that can do that decently well. I think I really need to work on increasing the distance between characters and deskewing the image.

For the distance between character's I think that morphological changes and upscaling will work the best.

I also need to rerun the accuracy tests.