File management
I looked at the results for upscaling, and Standard & Dyslexia had a fractional increase, but Handwritten actually went down a few precents.

Next I need to try to apply morphs to all of them. I will go with 3x3 kernel for the time being as it seems to have worked the best in the past