reorganized files
I'm going to try a new method to manipulate the fonts. I'm going to start with a control image, and then modify tat image.
For wieght, I might be able to apply some morphology from openCV.
Slant might be more difficult, I think that converting it to a bitmap or matrix might work by allowing me to perform a transformation on the actual pixels.
I think I will ignore width for now and instead focus on curviness, this will be the most difficult, I think I can find contours with openCV and store those values to manipulate the edges like with slant.